Move tests for Http\Client.
The code was moved in 3.3 and the tests were left alone to ensure
backwards compat. Now that development on the client has completed,
moving the tests makes sense.
