+++ tests/TestCase/Http/Client/CookieCollectionTest.php
@@ -11,7 +11,7 @@
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Test\TestCase\Network\Http;
+namespace Cake\Test\TestCase\Http\Client;
 
 use Cake\Http\Client\CookieCollection;
 use Cake\Http\Client\Response;
